requirements.txt
tensorflow==2.15.0
pillow
SimpleITK
imagecodecs
numpy
pandas
h5py